feat(FE): 인증 흐름 리팩토링 및 라우팅 개선
- 세션 스토리지를 사용하여 로그인 상태를 관리하는 사용자 인증 체크 구현
- App 컴포넌트를 업데이트하여 `useLoaderData`로 인증 상태 활용
- 인증되지 않은 사용자를 위해 AuthError 컴포넌트를 표시하도록 LoginRequiredRoute 리팩토링
- 404 페이지를 위한 새로운 NotFoundError 컴포넌트로 라우팅 개선
- Feed, WriteReview, Onboarding을 포함한 다양한 컴포넌트에 대한 새로운 라우트 통합
- 인증 상태에 따라 조건부로 렌더링되는 Header 컴포넌트 개선
- 라우트 및 로더에 대한 누락된 오류 처리 추가
