+++ front/src/pages/AuthError.tsx
@@ -1,0 +1,12 @@
+import KakaoLoginButton from "@/features/auth/KakaoLoginButton";
+
+export default function AuthError() {
+  return (
+    <div className="flex flex-col items-center justify-center h-screen">
+      <p className="text-2xl font-bold mb-10">
+        로그인 후 사용 가능한 기능입니다. 로그인 해주세요.
+      </p>
+      <KakaoLoginButton />
+    </div>
+  );
+}
